refactor: relocate config reader (global scope)
